+++ src/libs/animals/src/t_seagulls.cpp
@@ -43,11 +43,11 @@ void TSeagulls::LoadSettings()
     maxAngleSpeed = ini->GetFloat(ANIMALS_SEAGULLS_SECTION, "angle speed", SEAGULL_MAX_SPEED);
     maxDistance = ini->GetFloat(ANIMALS_SEAGULLS_SECTION, "distance", SEAGULL_DISTANCE);
     maxHeight = ini->GetFloat(ANIMALS_SEAGULLS_SECTION, "height", SEAGULL_MAX_HEIGHT);
-    maxCircleTime = ini->GetLong(ANIMALS_SEAGULLS_SECTION, "circle time", SEAGULL_MAX_CIRCLE_TIME);
-    farChoiceChance = ini->GetLong(ANIMALS_SEAGULLS_SECTION, "far choice", SEAGULL_LONG_DISTANCE_CHANCE);
-    relaxTime = ini->GetLong(ANIMALS_SEAGULLS_SECTION, "relax time", SEAGULL_RELAX_TIME);
-    screamTime = ini->GetLong(ANIMALS_SEAGULLS_SECTION, "scream time", SEAGULL_SCREAM_TIME);
-    countAdd = ini->GetLong(ANIMALS_SEAGULLS_SECTION, "add count", SEAGULL_ADD_COUNT);
+    maxCircleTime = ini->GetInt(ANIMALS_SEAGULLS_SECTION, "circle time", SEAGULL_MAX_CIRCLE_TIME);
+    farChoiceChance = ini->GetInt(ANIMALS_SEAGULLS_SECTION, "far choice", SEAGULL_LONG_DISTANCE_CHANCE);
+    relaxTime = ini->GetInt(ANIMALS_SEAGULLS_SECTION, "relax time", SEAGULL_RELAX_TIME);
+    screamTime = ini->GetInt(ANIMALS_SEAGULLS_SECTION, "scream time", SEAGULL_SCREAM_TIME);
+    countAdd = ini->GetInt(ANIMALS_SEAGULLS_SECTION, "add count", SEAGULL_ADD_COUNT);
     ini->ReadString(ANIMALS_SEAGULLS_SECTION, "scream file", screamFilename, 256, ANIMALS_SEAGULLS_SCREAM_FILENAME);
 }
