+++ src/libs/battle_interface/src/sea/battle_ship_command.cpp
@@ -161,7 +161,7 @@ int32_t BIShipCommandList::ShipAdding(bool allLabel, bool bMyShip, bool bEnemy, 
                 {
                     auto *pvdat =
                         core.Event("evntCheckEnableShip", "sl", m_sCurrentCommandName.c_str(), sd->characterIndex);
-                    if (pvdat != nullptr && pvdat->GetLong() == 0)
+                    if (pvdat != nullptr && pvdat->GetInt() == 0)
                         continue;
                 }
                 n = AddToIconList(sd->textureNum, sd->pictureNum, sd->selectPictureNum, -1, sd->characterIndex, nullptr,
@@ -218,7 +218,7 @@ int32_t BIShipCommandList::FortAdding(bool allLabel, bool bFriend, bool bNeutral
                 if (SQR(pL->x - selX) + SQR(pL->z - selZ) > sqrRadius)
                     continue;
             auto *pvdat = core.Event("evntCheckEnableLocator", "sa", m_sCurrentCommandName.c_str(), pL->pA);
-            if (pvdat != nullptr && pvdat->GetLong() == 0)
+            if (pvdat != nullptr && pvdat->GetInt() == 0)
                 continue;
             char *pLocName = nullptr;
             if (pL->pA != nullptr)
@@ -270,7 +270,7 @@ int32_t BIShipCommandList::LandAdding(bool allLabel)
             if (SQR(pL->x - selX) + SQR(pL->z - selZ) > sqrRadius)
                 continue;
         auto *pvdat = core.Event("evntCheckEnableLocator", "sa", m_sCurrentCommandName.c_str(), pL->pA);
-        if (pvdat != nullptr && pvdat->GetLong() == 0)
+        if (pvdat != nullptr && pvdat->GetInt() == 0)
             continue;
         char *pLocName = nullptr;
         if (pL->pA != nullptr)
@@ -460,7 +460,7 @@ int32_t BIShipCommandList::TownAdding(bool allLabel, bool bDiseased, bool bNotDi
             if (SQR(pL->x - selX) + SQR(pL->z - selZ) > sqrRadius)
                 continue;
         auto *pvdat = core.Event("evntCheckEnableLocator", "sa", m_sCurrentCommandName.c_str(), pL->pA);
-        if (pvdat != nullptr && pvdat->GetLong() == 0)
+        if (pvdat != nullptr && pvdat->GetInt() == 0)
             continue;
         char *pLocName = nullptr;
         if (pL->pA != nullptr)
