+++ src/libs/battle_interface/src/sea/battle_ship_sign.cpp
@@ -784,5 +784,5 @@ int32_t BIShipIcon::GetShipClass(int32_t nCharIdx)
     VDATA *pVDat = core.Event("evntGetCharacterShipClass", "l", nCharIdx);
     if (!pVDat)
         return 1;
-    return pVDat->GetLong();
+    return pVDat->GetInt();
 }
