+++ src/libs/common/include/weather_base.h
@@ -88,7 +88,7 @@ class WEATHER_BASE : public Entity
 
     ~WEATHER_BASE() override{};
 
-    virtual int32_t GetLong(uint32_t dwCode) = 0;
+    virtual int32_t GetInt(uint32_t dwCode) = 0;
     virtual uint32_t GetColor(uint32_t, CVECTOR *) = 0;
     virtual uint32_t GetColor(uint32_t) = 0;
     virtual float GetFloat(uint32_t) = 0;
