+++ src/libs/lighter/src/lighter.cpp
@@ -39,11 +39,11 @@ bool Lighter::Init()
     auto ini = fio->OpenIniFile("resource\\ini\\loclighter.ini");
     if (!ini)
         return false;
-    const auto isLoading = ini->GetLong(nullptr, "loading", 0);
-    autoTrace = ini->GetLong(nullptr, "autotrace", 0) != 0;
-    autoSmooth = ini->GetLong(nullptr, "autosmooth", 0) != 0;
-    window.isSmallSlider = ini->GetLong(nullptr, "smallslider", 0) != 0;
-    geometry.useColor = ini->GetLong(nullptr, "usecolor", 0) != 0;
+    const auto isLoading = ini->GetInt(nullptr, "loading", 0);
+    autoTrace = ini->GetInt(nullptr, "autotrace", 0) != 0;
+    autoSmooth = ini->GetInt(nullptr, "autosmooth", 0) != 0;
+    window.isSmallSlider = ini->GetInt(nullptr, "smallslider", 0) != 0;
+    geometry.useColor = ini->GetInt(nullptr, "usecolor", 0) != 0;
     if (!isLoading)
         return false;
     // DX9 render
